Update KR_United_States_of_America_l_english.yml
Hey guys, i'm new here but I thought I would try and contribute a little bit to this project so I fixed a bunch of typos in the USA localization and fixed some of the focuses too. Particularly, there were focuses that had tooltips with "modifier_experience_gain" instead of saying "Motorized Training Experience Gain: -5%" so I fixed that. If there's something wrong with it i'd be happy to go back and reverse the changes I made but I tested it and everything looks good.